post_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pyjwt
flask-restx
prometheus_flask_exporter
marshmallow
Werkzeug
flask_cors
pytest
pytest-flask
requests
